Migrate and merge otlp utils to core.
