+++ newrelic/core/agent_protocol.py
@@ -46,7 +46,7 @@ from newrelic.network.exceptions import (
     NetworkInterfaceException,
     RetryDataForRequest,
 )
-from newrelic.common.otlp_utils import OTLP_CONTENT_TYPE, otlp_encode
+from newrelic.core.otlp_utils import OTLP_CONTENT_TYPE, otlp_encode
 
 _logger = logging.getLogger(__name__)
